+++ x-pack/plugins/ml/public/plugin.ts
@@ -12,6 +12,8 @@ import {
   AppMountParameters,
   PluginInitializerContext,
 } from 'kibana/public';
+import { BehaviorSubject } from 'rxjs';
+import { take } from 'rxjs/operators';
 import { ManagementSetup } from 'src/plugins/management/public';
 import { SharePluginSetup, SharePluginStart, UrlGeneratorState } from 'src/plugins/share/public';
 import { UsageCollectionSetup } from 'src/plugins/usage_collection/server';
@@ -19,9 +21,10 @@ import { UsageCollectionSetup } from 'src/plugins/usage_collection/server';
 import { DataPublicPluginStart } from 'src/plugins/data/public';
 import { HomePublicPluginSetup } from 'src/plugins/home/public';
 import { EmbeddableSetup } from 'src/plugins/embeddable/public';
+import { AppStatus, AppUpdater } from '../../../../src/core/public';
 import { SecurityPluginSetup } from '../../security/public';
 import { LicensingPluginSetup } from '../../licensing/public';
-import { initManagementSection } from './application/management';
+import { registerManagementSection } from './application/management';
 import { LicenseManagementUIPluginSetup } from '../../license_management/public';
 import { setDependencyCache } from './application/util/dependency_cache';
 import { PLUGIN_ID, PLUGIN_ICON } from '../common/constants/app';
@@ -31,7 +34,8 @@ import { registerEmbeddables } from './embeddables';
 import { UiActionsSetup, UiActionsStart } from '../../../../src/plugins/ui_actions/public';
 import { registerMlUiActions } from './ui_actions';
 import { KibanaLegacyStart } from '../../../../src/plugins/kibana_legacy/public';
-import { MlUrlGenerator, MlUrlGeneratorState, ML_APP_URL_GENERATOR } from './url_generator';
+import { registerUrlGenerator, MlUrlGeneratorState, ML_APP_URL_GENERATOR } from './url_generator';
+import { isMlEnabled, isFullLicense } from '../common/license';
 
 export interface MlStartDependencies {
   data: DataPublicPluginStart;
@@ -61,18 +65,11 @@ declare module '../../../../src/plugins/share/public' {
 export type MlCoreSetup = CoreSetup<MlStartDependencies, MlPluginStart>;
 
 export class MlPlugin implements Plugin<MlPluginSetup, MlPluginStart> {
+  private appUpdater = new BehaviorSubject<AppUpdater>(() => ({}));
+
   constructor(private initializerContext: PluginInitializerContext) {}
 
   setup(core: MlCoreSetup, pluginsSetup: MlSetupDependencies) {
-    const baseUrl = core.http.basePath.prepend('/app/ml');
-
-    pluginsSetup.share.urlGenerators.registerUrlGenerator(
-      new MlUrlGenerator({
-        appBasePath: baseUrl,
-        useHash: core.uiSettings.get('state:storeInSessionStorage'),
-      })
-    );
-
     core.application.register({
       id: PLUGIN_ID,
       title: i18n.translate('xpack.ml.plugin.title', {
@@ -82,6 +79,7 @@ export class MlPlugin implements Plugin<MlPluginSetup, MlPluginStart> {
       euiIconType: PLUGIN_ICON,
       appRoute: '/app/ml',
       category: DEFAULT_APP_CATEGORIES.kibana,
+      updater$: this.appUpdater,
       mount: async (params: AppMountParameters) => {
         const [coreStart, pluginsStart] = await core.getStartServices();
         const kibanaVersion = this.initializerContext.env.packageInfo.version;
@@ -112,11 +110,26 @@ export class MlPlugin implements Plugin<MlPluginSetup, MlPluginStart> {
       },
     });
 
-    registerFeature(pluginsSetup.home);
+    const licensing = pluginsSetup.licensing.license$.pipe(take(1));
+    licensing.subscribe((license) => {
+      if (isMlEnabled(license)) {
+        // add ML to home page
+        registerFeature(pluginsSetup.home);
 
-    initManagementSection(pluginsSetup, core);
-    registerEmbeddables(pluginsSetup.embeddable, core);
-    registerMlUiActions(pluginsSetup.uiActions, core);
+        // register various ML plugin features which require a full license
+        if (isFullLicense(license)) {
+          registerManagementSection(pluginsSetup.management, core);
+          registerEmbeddables(pluginsSetup.embeddable, core);
+          registerMlUiActions(pluginsSetup.uiActions, core);
+          registerUrlGenerator(pluginsSetup.share, core);
+        }
+      } else {
+        // if ml is disabled in elasticsearch, disable ML in kibana
+        this.appUpdater.next(() => ({
+          status: AppStatus.inaccessible,
+        }));
+      }
+    });
 
     return {};
   }
